Moved datatypes to their own module
