+++ yml2block/tsv_input.py
@@ -7,71 +7,7 @@ import io
 import itertools
 
 from yml2block.rules import LintViolation, Level
-
-
-class MDBlockList(list):
-    __slots__ = ("line", "column")
-
-    def __init__(self, iterable=None, line=None, column=None):
-        """Create a new MDBlock list that behaves like a normal list
-        but tracks line an column number in object slots.
-        """
-        self.line = line
-        self.column = column
-
-        # Delegate initialization to the list constructore
-        if iterable is None:
-            super().__init__()
-        else:
-            super().__init__(iterable)
-
-    @classmethod
-    def from_ruamel(cls, ruamel_list):
-        """Create a MDBlock list from an existing list read by the
-        ruamel.yaml round trip parser.
-        """
-        return cls(ruamel_list, line=ruamel_list.lc.line, column=ruamel_list.lc.col)
-
-
-class MDBlockDict(dict):
-    __slots__ = ("line", "column")
-
-    def __init__(self, mapping=None, /, line=None, column=None, **kwargs):
-        """Create a new MDBlock Dictionary that behaves like a normal dictionary
-        but tracks line an column number in object slots.
-        """
-        self.line = line
-        self.column = column
-
-        # Ensure an inner dict exists
-        if mapping is None:
-            mapping = dict()
-
-        # Remain compatible with pythons regular dict constructor by
-        # allowing kwargs to extend the passed mapping
-        mapping.update(kwargs)
-
-        super().__init__(mapping)
-
-    @classmethod
-    def from_ruamel(cls, ruamel_dict):
-        """Create a MDBlock dict from an existing dict read by the
-        ruamel.yaml round trip parser.
-        """
-        return cls(ruamel_dict, line=ruamel_dict.lc.line, column=ruamel_dict.lc.col)
-
-
-class MDBlockNode:
-    __slots__ = ("line", "column", "value")
-
-    def __init__(self, value, line=None, column=None):
-        self.line = line
-        self.column = column
-        self.value = value
-
-    def __repr__(self):
-        return f"({self.line}, {self.column}) {self.value}"
-
+from yml2block.datatypes import MDBlockList, MDBlockDict, MDBlockNode
 
 def _identify_break_points(full_file):
     """Identify where to split the metadata block into its three subsections"""
